+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterAdValue.java
@@ -12,5 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/** Versioned request agent string. */
-#define FLT_REQUEST_AGENT_VERSIONED @"Flutter-GMA-0.13.2"
+package io.flutter.plugins.googlemobileads;
+
+import androidx.annotation.NonNull;
+
+/** A wrapper for {@link com.google.android.gms.ads.AdValue}. */
+public class FlutterAdValue {
+  final int precisionType;
+  @NonNull final String currencyCode;
+  final long valueMicros;
+
+  public FlutterAdValue(int precisionType, @NonNull String currencyCode, long valueMicros) {
+    this.precisionType = precisionType;
+    this.currencyCode = currencyCode;
+    this.valueMicros = valueMicros;
+  }
+}