+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterFullScreenContentCallback.java
@@ -25,31 +25,31 @@ import com.google.android.gms.ads.FullScreenContentCallback;
 class FlutterFullScreenContentCallback extends FullScreenContentCallback {
 
   @NonNull protected final AdInstanceManager manager;
-  @NonNull protected final FlutterAd ad;
 
-  public FlutterFullScreenContentCallback(
-      @NonNull AdInstanceManager manager, @NonNull FlutterAd ad) {
+  @NonNull protected final int adId;
+
+  public FlutterFullScreenContentCallback(@NonNull AdInstanceManager manager, int adId) {
     this.manager = manager;
-    this.ad = ad;
+    this.adId = adId;
   }
 
   @Override
   public void onAdFailedToShowFullScreenContent(@NonNull AdError adError) {
-    manager.onFailedToShowFullScreenContent(ad, adError);
+    manager.onFailedToShowFullScreenContent(adId, adError);
   }
 
   @Override
   public void onAdShowedFullScreenContent() {
-    manager.onAdShowedFullScreenContent(ad);
+    manager.onAdShowedFullScreenContent(adId);
   }
 
   @Override
   public void onAdDismissedFullScreenContent() {
-    manager.onAdDismissedFullScreenContent(ad);
+    manager.onAdDismissedFullScreenContent(adId);
   }
 
   @Override
   public void onAdImpression() {
-    manager.onAdImpression(ad);
+    manager.onAdImpression(adId);
   }
 }