+++ packages/google_mobile_ads/android/src/test/java/io/flutter/plugins/googlemobileads/FlutterRewardedAdTest.java
@@ -14,6 +14,9 @@
 
 package io.flutter.plugins.googlemobileads;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
+import static org.mockito.ArgumentCaptor.forClass;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
@@ -29,6 +32,7 @@ import android.content.Context;
 import androidx.annotation.Nullable;
 import com.google.android.gms.ads.AdError;
 import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.AdValue;
 import com.google.android.gms.ads.FullScreenContentCallback;
 import com.google.android.gms.ads.LoadAdError;
 import com.google.android.gms.ads.OnUserEarnedRewardListener;
@@ -44,6 +48,7 @@ import io.flutter.plugins.googlemobileads.FlutterAd.FlutterLoadAdError;
 import io.flutter.plugins.googlemobileads.FlutterRewardedAd.FlutterRewardItem;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.ArgumentCaptor;
 import org.mockito.ArgumentMatcher;
 import org.mockito.ArgumentMatchers;
 import org.mockito.invocation.InvocationOnMock;
@@ -72,7 +77,7 @@ public class FlutterRewardedAdTest {
     when(mockFlutterAdRequest.asAdRequest()).thenReturn(mockAdRequest);
     flutterRewardedAd =
         new FlutterRewardedAd(
-            mockManager, "testId", mockFlutterAdRequest, options, mockFlutterAdLoader);
+            1, mockManager, "testId", mockFlutterAdRequest, options, mockFlutterAdLoader);
   }
 
   private void setupAdManagerMocks(@Nullable FlutterServerSideVerificationOptions options) {
@@ -81,7 +86,7 @@ public class FlutterRewardedAdTest {
     when(mockAdManagerFlutterRequest.asAdManagerAdRequest()).thenReturn(mockAdManagerAdRequest);
     flutterRewardedAd =
         new FlutterRewardedAd(
-            mockManager, "testId", mockAdManagerFlutterRequest, options, mockFlutterAdLoader);
+            1, mockManager, "testId", mockAdManagerFlutterRequest, options, mockFlutterAdLoader);
   }
 
   @Test
@@ -119,7 +124,7 @@ public class FlutterRewardedAdTest {
             any(RewardedAdLoadCallback.class));
 
     FlutterLoadAdError expectedError = new FlutterLoadAdError(loadAdError);
-    verify(mockManager).onAdFailedToLoad(eq(flutterRewardedAd), eq(expectedError));
+    verify(mockManager).onAdFailedToLoad(eq(1), eq(expectedError));
   }
 
   @Test
@@ -129,7 +134,6 @@ public class FlutterRewardedAdTest {
     setupAdManagerMocks(options);
 
     final RewardedAd mockAd = mock(RewardedAd.class);
-    final LoadAdError loadAdError = new LoadAdError(1, "2", "3", null, null);
     doAnswer(
             new Answer() {
               @Override
@@ -148,6 +152,23 @@ public class FlutterRewardedAdTest {
             any(RewardedAdLoadCallback.class));
     final ResponseInfo responseInfo = mock(ResponseInfo.class);
     doReturn(responseInfo).when(mockAd).getResponseInfo();
+
+    final AdValue adValue = mock(AdValue.class);
+    doReturn(1).when(adValue).getPrecisionType();
+    doReturn("Dollars").when(adValue).getCurrencyCode();
+    doReturn(1000L).when(adValue).getValueMicros();
+    doAnswer(
+            new Answer() {
+              @Override
+              public Object answer(InvocationOnMock invocation) {
+                FlutterPaidEventListener listener = invocation.getArgument(0);
+                listener.onPaidEvent(adValue);
+                return null;
+              }
+            })
+        .when(mockAd)
+        .setOnPaidEventListener(any(FlutterPaidEventListener.class));
+
     flutterRewardedAd.load();
 
     verify(mockFlutterAdLoader)
@@ -157,8 +178,15 @@ public class FlutterRewardedAdTest {
             eq(mockAdManagerAdRequest),
             any(RewardedAdLoadCallback.class));
 
-    verify(mockManager).onAdLoaded(eq(flutterRewardedAd), eq(responseInfo));
+    verify(mockManager).onAdLoaded(eq(1), eq(responseInfo));
+    verify(mockAd).setOnPaidEventListener(any(FlutterPaidEventListener.class));
+    final ArgumentCaptor<FlutterAdValue> adValueCaptor = forClass(FlutterAdValue.class);
+    verify(mockManager).onPaidEvent(eq(flutterRewardedAd), adValueCaptor.capture());
+    assertEquals(adValueCaptor.getValue().currencyCode, "Dollars");
+    assertEquals(adValueCaptor.getValue().precisionType, 1);
+    assertEquals(adValueCaptor.getValue().valueMicros, 1000L);
 
+    // Setup mocks for show().
     doAnswer(
             new Answer() {
               @Override
@@ -215,12 +243,13 @@ public class FlutterRewardedAdTest {
     verify(mockAd)
         .setServerSideVerificationOptions(
             ArgumentMatchers.argThat(serverSideVerificationOptionsArgumentMatcher));
-    verify(mockManager).onAdShowedFullScreenContent(eq(flutterRewardedAd));
-    verify(mockManager).onAdImpression(eq(flutterRewardedAd));
-    verify(mockManager).onAdDismissedFullScreenContent(eq(flutterRewardedAd));
-    verify(mockManager)
-        .onRewardedAdUserEarnedReward(flutterRewardedAd, new FlutterRewardItem(5, "$$"));
-    verify(mockManager).onAdMetadataChanged(eq(flutterRewardedAd));
+    verify(mockManager).onAdShowedFullScreenContent(eq(1));
+    verify(mockManager).onAdImpression(eq(1));
+    verify(mockManager).onAdDismissedFullScreenContent(eq(1));
+    verify(mockManager).onRewardedAdUserEarnedReward(1, new FlutterRewardItem(5, "$$"));
+    verify(mockManager).onAdMetadataChanged(eq(1));
+
+    assertNull(flutterRewardedAd.getPlatformView());
   }
 
   @Test
@@ -260,7 +289,7 @@ public class FlutterRewardedAdTest {
             eq(mockAdManagerAdRequest),
             any(RewardedAdLoadCallback.class));
 
-    verify(mockManager).onAdLoaded(eq(mockFlutterAd), eq(responseInfo));
+    verify(mockManager).onAdLoaded(eq(1), eq(responseInfo));
 
     doAnswer(
             new Answer() {
@@ -322,7 +351,7 @@ public class FlutterRewardedAdTest {
             eq(mockAdRequest),
             any(RewardedAdLoadCallback.class));
 
-    verify(mockManager).onAdLoaded(eq(flutterRewardedAd), eq(responseInfo));
+    verify(mockManager).onAdLoaded(eq(1), eq(responseInfo));
     final AdError adError = new AdError(0, "ad", "error");
     doAnswer(
             new Answer() {
@@ -338,6 +367,6 @@ public class FlutterRewardedAdTest {
 
     flutterRewardedAd.show();
     verify(mockRewardedAd).setFullScreenContentCallback(any(FullScreenContentCallback.class));
-    verify(mockManager).onFailedToShowFullScreenContent(eq(flutterRewardedAd), eq(adError));
+    verify(mockManager).onFailedToShowFullScreenContent(eq(1), eq(adError));
   }
 }