+++ packages/google_mobile_ads/ios/Tests/FLTInterstitialAdTest.m
@@ -38,7 +38,8 @@
   UIViewController *mockRootViewController = OCMClassMock([UIViewController class]);
   FLTInterstitialAd *ad = [[FLTInterstitialAd alloc] initWithAdUnitId:@"testId"
                                                               request:request
-                                                   rootViewController:mockRootViewController];
+                                                   rootViewController:mockRootViewController
+                                                                 adId:@1];
   ad.manager = mockManager;
 
   id interstitialClassMock = OCMClassMock([GADInterstitialAd class]);
@@ -66,6 +67,17 @@
   GADResponseInfo *mockResponseInfo = OCMClassMock([GADResponseInfo class]);
   OCMStub([interstitialClassMock responseInfo]).andReturn(mockResponseInfo);
 
+  // Mock callback of paid event handler.
+  GADAdValue *adValue = OCMClassMock([GADAdValue class]);
+  OCMStub([adValue value]).andReturn(NSDecimalNumber.one);
+  OCMStub([adValue precision]).andReturn(GADAdValuePrecisionEstimated);
+  OCMStub([adValue currencyCode]).andReturn(@"currencyCode");
+  OCMStub([interstitialClassMock setPaidEventHandler:[OCMArg checkWithBlock:^BOOL(id obj) {
+                                   GADPaidEventHandler handler = obj;
+                                   handler(adValue);
+                                   return YES;
+                                 }]]);
+  // Call load and verify interactions with mocks.
   [ad load];
 
   OCMVerify(ClassMethod([interstitialClassMock loadWithAdUnitID:[OCMArg isEqual:@"testId"]
@@ -75,6 +87,16 @@
                        responseInfo:[OCMArg isEqual:mockResponseInfo]]);
   OCMVerify([interstitialClassMock setFullScreenContentDelegate:[OCMArg isEqual:ad]]);
   XCTAssertEqual(ad.interstitial, interstitialClassMock);
+  OCMVerify([mockManager onPaidEvent:[OCMArg isEqual:ad]
+                               value:[OCMArg checkWithBlock:^BOOL(id obj) {
+                                 FLTAdValue *adValue = obj;
+                                 XCTAssertEqualObjects(
+                                     adValue.valueMicros,
+                                     [[NSDecimalNumber alloc] initWithInt:1000000]);
+                                 XCTAssertEqual(adValue.precision, GADAdValuePrecisionEstimated);
+                                 XCTAssertEqualObjects(adValue.currencyCode, @"currencyCode");
+                                 return TRUE;
+                               }]]);
 
   // Show the ad
   [ad show];
@@ -100,7 +122,8 @@
   UIViewController *mockRootViewController = OCMClassMock([UIViewController class]);
   FLTInterstitialAd *ad = [[FLTInterstitialAd alloc] initWithAdUnitId:@"testId"
                                                               request:request
-                                                   rootViewController:mockRootViewController];
+                                                   rootViewController:mockRootViewController
+                                                                 adId:@1];
   ad.manager = mockManager;
 
   id interstitialClassMock = OCMClassMock([GADInterstitialAd class]);